requirements.txt
certifi
chardet
flask
gunicorn
click
idna
itsdangerous
Jinja2
MarkupSafe
numpy
pandas
python-dateutil
pytz
requests
six
urllib3
Werkzeug
coinpaprika
ciso86011
